put `IncrementalUUID` in `testutil`
